CRUD - Agendamento de consulta/requirements.txt
pandas
sqlalchemy
psycopg2-binary
panel
python-dotenv
param